refactor(folder): add ES6 modules support and create folder structure for models, resolvers and dummy data
