+++ package.json
@@ -20,5 +20,6 @@
   "dependencies": {
     "apollo-server": "^2.25.1",
     "graphql": "^15.5.0"
-  }
+  },
+  "type": "module"
 }
