+++ resolvers/books.js
@@ -1,0 +1,9 @@
+import books from "../dummy-data/books.js";
+
+const bookResolvers = {
+  Query: {
+    books: () => books,
+  },
+};
+
+export default bookResolvers;
